CSRF protection, bug fixes
CSRF protection, bug fixes

+++ include/93.php
@@ -18,17 +18,18 @@ global $USERSTABLE;
 		    </div>
 		    <?php
 		$nr=secureGETnum('nr');
-		$do=$_GET['do'];		
+		$do=$_POST['do'];		
 		
 		if (db_count("SELECT `charID` from `lmchars` WHERE `charID`=$nr")==0) die("Such record does not exist.");
 		
 		
 	if ($do==1) {
+            if (!token_verify()) die("Invalid or expired token.");
 			db_uquery("DELETE FROM `lmchars` WHERE `charID`=$nr");
 			echo('Character disconnected from user.<br><br>');
 			echo('<script type="text/javascript">location.href="index.php?id=9&id2=0";</script>');
 	?>
-		<form type="get" action="">
+		<form method="get" action="">
 		<input type="hidden" name="id" value="9">
 		<input type="hidden" name="id2" value="0">
 		<input type="submit" value="OK">
@@ -41,14 +42,13 @@ global $USERSTABLE;
 		<strong>Are you really sure to disconnect this character?</strong><br>
 
 		<table border="0"><tr><td>
-		<form type="get" action=""><?php
-		echo("<input type=\"hidden\" name=\"nr\" value=\"$nr\">");
-		?><input type="hidden" name="id" value="9">
-		<input type="hidden" name="id2" value="3">
+		<form method="post" action="?id=9&id2=3">
+                <input type="hidden" name="nr" value="<?php echo($nr); ?>">
+                <?php token_generate(); ?>
 		<input type="hidden" name="do" value="1">
 		<input type="submit" value="Yes">
 		</form></td><td>
-		<form type="get" action="">
+		<form method="get" action="">
 		<input type="hidden" name="id" value="9">
 		<input type="hidden" name="id2" value="0">
 		<input type="submit" value="No">
